chore: removed NetInfo.tsx and moved functionality to network.tsx
